+++ io7m-jcanephora-jogl/src/main/java/com/io7m/jcanephora/jogl/JCGLInterfaceGLES3_JOGL_ES3.java
@@ -1,10 +1,10 @@
 /*
  * Copyright © 2014 <code@io7m.com> http://io7m.com
- * 
+ *
  * Permission to use, copy, modify, and/or distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
  * copyright notice and this permission notice appear in all copies.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
  * WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
  * MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY
@@ -919,7 +919,7 @@ final class JCGLInterfaceGLES3_JOGL_ES3 implements JCGLInterfaceGLES3Type
 
   @Override public IndexBufferType indexBufferAllocate(
     final ArrayBufferUsableType buffer,
-    final int indices,
+    final long indices,
     final UsageHint usage)
     throws JCGLException
   {
@@ -928,7 +928,7 @@ final class JCGLInterfaceGLES3_JOGL_ES3 implements JCGLInterfaceGLES3Type
 
   @Override public IndexBufferType indexBufferAllocateType(
     final JCGLUnsignedType type,
-    final int indices,
+    final long indices,
     final UsageHint usage)
     throws JCGLException
   {
